@@ -1,32 +1,27 @@
 {
-  "_args": [
-    [
-      "cidr-regex@2.0.9",
-      "/Users/rebecca/code/npm"
-    ]
-  ],
-  "_from": "cidr-regex@2.0.9",
-  "_id": "cidr-regex@2.0.9",
+  "_from": "cidr-regex@^2.0.10",
+  "_id": "cidr-regex@2.0.10",
   "_inBundle": true,
-  "_integrity": "sha512-F7/fBRUU45FnvSPjXdpIrc++WRSBdCiSTlyq4ZNhLKOlHFNWgtzZ0Fd+zrqI/J1j0wmlx/f5ZQDmD2GcbrNcmw==",
+  "_integrity": "sha512-sB3ogMQXWvreNPbJUZMRApxuRYd+KoIo4RGQ81VatjmMW6WJPo+IJZ2846FGItr9VzKo5w7DXzijPLGtSd0N3Q==",
   "_location": "/npm/cidr-regex",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "range",
     "registry": true,
-    "raw": "cidr-regex@2.0.9",
+    "raw": "cidr-regex@^2.0.10",
     "name": "cidr-regex",
     "escapedName": "cidr-regex",
-    "rawSpec": "2.0.9",
+    "rawSpec": "^2.0.10",
     "saveSpec": null,
-    "fetchSpec": "2.0.9"
+    "fetchSpec": "^2.0.10"
   },
   "_requiredBy": [
     "/npm/is-cidr"
   ],
-  "_resolved": "https://registry.npmjs.org/cidr-regex/-/cidr-regex-2.0.9.tgz",
-  "_spec": "2.0.9",
-  "_where": "/Users/rebecca/code/npm",
+  "_resolved": "https://registry.npmjs.org/cidr-regex/-/cidr-regex-2.0.10.tgz",
+  "_shasum": "af13878bd4ad704de77d6dc800799358b3afa70d",
+  "_spec": "cidr-regex@^2.0.10",
+  "_where": "/Users/aeschright/code/cli/node_modules/is-cidr",
   "author": {
     "name": "silverwind",
     "email": "me@silverwind.io"
@@ -34,6 +29,7 @@
   "bugs": {
     "url": "https://github.com/silverwind/cidr-regex/issues"
   },
+  "bundleDependencies": false,
   "contributors": [
     {
       "name": "Felipe Apostol",
@@ -44,12 +40,13 @@
   "dependencies": {
     "ip-regex": "^2.1.0"
   },
+  "deprecated": false,
   "description": "Regular expression for matching IP addresses in CIDR notation",
   "devDependencies": {
-    "ava": "^0.25.0",
-    "eslint": "^4.19.1",
-    "eslint-config-silverwind": "^1.0.42",
-    "updates": "^3.0.0"
+    "eslint": "^5.6.0",
+    "eslint-config-silverwind": "^2.0.9",
+    "updates": "^4.3.0",
+    "ver": "^2.0.1"
   },
   "engines": {
     "node": ">=4"
@@ -77,5 +74,5 @@
   "scripts": {
     "test": "make test"
   },
-  "version": "2.0.9"
+  "version": "2.0.10"
 }
