@@ -1,29 +1,33 @@
 {
-  "_from": "get-stream@^3.0.0",
-  "_id": "get-stream@3.0.0",
+  "_from": "get-stream@",
+  "_id": "get-stream@4.1.0",
   "_inBundle": true,
-  "_integrity": "sha1-jpQ9E1jcN1VQVOy+LtsFqhdO3hQ=",
+  "_integrity": "sha512-GMat4EJ5161kIy2HevLlr4luNjBgvmj413KaQA7jt4V8B4RDsfpHk7WQ9GVqfYyyx8OS/L66Kox+rJRNklLK7w==",
   "_location": "/npm/get-stream",
   "_phantomChildren": {},
   "_requested": {
-    "type": "range",
+    "type": "tag",
     "registry": true,
-    "raw": "get-stream@^3.0.0",
+    "raw": "get-stream@",
     "name": "get-stream",
     "escapedName": "get-stream",
-    "rawSpec": "^3.0.0",
+    "rawSpec": "",
     "saveSpec": null,
-    "fetchSpec": "^3.0.0"
+    "fetchSpec": "latest"
   },
   "_requiredBy": [
-    "/npm/execa",
-    "/npm/got",
+    "/npm/libnpmaccess",
+    "/npm/libnpmhook",
+    "/npm/libnpmorg",
+    "/npm/libnpmpublish",
+    "/npm/libnpmsearch",
+    "/npm/libnpmteam",
     "/npm/pacote"
   ],
-  "_resolved": "https://registry.npmjs.org/get-stream/-/get-stream-3.0.0.tgz",
-  "_shasum": "8e943d1358dc37555054ecbe2edb05aa174ede14",
-  "_spec": "get-stream@^3.0.0",
-  "_where": "/Users/rebecca/code/npm/node_modules/pacote",
+  "_resolved": "https://registry.npmjs.org/get-stream/-/get-stream-4.1.0.tgz",
+  "_shasum": "c1b255575f3dc21d59bfc79cd3d2b46b1c3a54b5",
+  "_spec": "get-stream@",
+  "_where": "/Users/zkat/Documents/code/work/npm",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -33,6 +37,9 @@
     "url": "https://github.com/sindresorhus/get-stream/issues"
   },
   "bundleDependencies": false,
+  "dependencies": {
+    "pump": "^3.0.0"
+  },
   "deprecated": false,
   "description": "Get a stream as a string, buffer, or array",
   "devDependencies": {
@@ -41,7 +48,7 @@
     "xo": "*"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=6"
   },
   "files": [
     "index.js",
@@ -54,7 +61,6 @@
     "promise",
     "concat",
     "string",
-    "str",
     "text",
     "buffer",
     "read",
@@ -63,8 +69,7 @@
     "readable",
     "readablestream",
     "array",
-    "object",
-    "obj"
+    "object"
   ],
   "license": "MIT",
   "name": "get-stream",
@@ -75,8 +80,5 @@
   "scripts": {
     "test": "xo && ava"
   },
-  "version": "3.0.0",
-  "xo": {
-    "esnext": true
-  }
+  "version": "4.1.0"
 }
