@@ -1,27 +1,27 @@
 {
-  "_from": "is-cidr@2.0.6",
-  "_id": "is-cidr@2.0.6",
+  "_from": "is-cidr@3.0.0",
+  "_id": "is-cidr@3.0.0",
   "_inBundle": true,
-  "_integrity": "sha512-A578p1dV22TgPXn6NCaDAPj6vJvYsBgAzUrAd28a4oldeXJjWqEUuSZOLIW3im51mazOKsoyVp8NU/OItlWacw==",
+  "_integrity": "sha512-8Xnnbjsb0x462VoYiGlhEi+drY8SFwrHiSYuzc/CEwco55vkehTaxAyIjEdpi3EMvLPPJAJi9FlzP+h+03gp0Q==",
   "_location": "/npm/is-cidr",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "is-cidr@2.0.6",
+    "raw": "is-cidr@3.0.0",
     "name": "is-cidr",
     "escapedName": "is-cidr",
-    "rawSpec": "2.0.6",
+    "rawSpec": "3.0.0",
     "saveSpec": null,
-    "fetchSpec": "2.0.6"
+    "fetchSpec": "3.0.0"
   },
   "_requiredBy": [
     "/npm"
   ],
-  "_resolved": "https://registry.npmjs.org/is-cidr/-/is-cidr-2.0.6.tgz",
-  "_shasum": "4b01c9693d8e18399dacd18a4f3d60ea5871ac60",
-  "_spec": "is-cidr@2.0.6",
-  "_where": "/Users/rebecca/code/npm",
+  "_resolved": "https://registry.npmjs.org/is-cidr/-/is-cidr-3.0.0.tgz",
+  "_shasum": "1acf35c9e881063cd5f696d48959b30fed3eed56",
+  "_spec": "is-cidr@3.0.0",
+  "_where": "/Users/aeschright/code/cli",
   "author": {
     "name": "silverwind",
     "email": "me@silverwind.io"
@@ -38,18 +38,18 @@
     }
   ],
   "dependencies": {
-    "cidr-regex": "^2.0.8"
+    "cidr-regex": "^2.0.10"
   },
   "deprecated": false,
   "description": "Check if a string is an IP address in CIDR notation",
   "devDependencies": {
-    "ava": "^0.25.0",
-    "eslint": "^4.19.1",
-    "eslint-config-silverwind": "^1.0.42",
-    "updates": "^3.0.0"
+    "eslint": "^5.7.0",
+    "eslint-config-silverwind": "^2.0.9",
+    "updates": "^4.5.2",
+    "ver": "^3.0.0"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=6"
   },
   "files": [
     "index.js"
@@ -63,7 +63,8 @@
     "prefix",
     "prefixes",
     "ip",
-    "ip address"
+    "ip address",
+    "network"
   ],
   "license": "BSD-2-Clause",
   "name": "is-cidr",
@@ -74,5 +75,5 @@
   "scripts": {
     "test": "make test"
   },
-  "version": "2.0.6"
+  "version": "3.0.0"
 }
