@@ -1,28 +1,28 @@
 {
-  "_from": "spdx-license-ids@^3.0.0",
-  "_id": "spdx-license-ids@3.0.0",
+  "_from": "spdx-license-ids@3.0.3",
+  "_id": "spdx-license-ids@3.0.3",
   "_inBundle": true,
-  "_integrity": "sha512-2+EPwgbnmOIl8HjGBXXMd9NAu02vLjOO1nWw4kmeRDFyHn+M/ETfHxQUK0oXg8ctgVnl9t3rosNVsZ1jG61nDA==",
+  "_integrity": "sha512-uBIcIl3Ih6Phe3XHK1NqboJLdGfwr1UN3k6wSD1dZpmPsIkb8AGNbZYJ1fOBk834+Gxy8rpfDxrS6XLEMZMY2g==",
   "_location": "/npm/spdx-license-ids",
   "_phantomChildren": {},
   "_requested": {
-    "type": "range",
+    "type": "version",
     "registry": true,
-    "raw": "spdx-license-ids@^3.0.0",
+    "raw": "spdx-license-ids@3.0.3",
     "name": "spdx-license-ids",
     "escapedName": "spdx-license-ids",
-    "rawSpec": "^3.0.0",
+    "rawSpec": "3.0.3",
     "saveSpec": null,
-    "fetchSpec": "^3.0.0"
+    "fetchSpec": "3.0.3"
   },
   "_requiredBy": [
     "/npm/spdx-correct",
     "/npm/spdx-expression-parse"
   ],
-  "_resolved": "https://registry.npmjs.org/spdx-license-ids/-/spdx-license-ids-3.0.0.tgz",
-  "_shasum": "7a7cd28470cc6d3a1cfe6d66886f6bc430d3ac87",
-  "_spec": "spdx-license-ids@^3.0.0",
-  "_where": "/Users/rebecca/code/npm/node_modules/spdx-correct",
+  "_resolved": "https://registry.npmjs.org/spdx-license-ids/-/spdx-license-ids-3.0.3.tgz",
+  "_shasum": "81c0ce8f21474756148bbb5f3bfc0f36bf15d76e",
+  "_spec": "spdx-license-ids@3.0.3",
+  "_where": "/Users/zkat/Documents/code/work/npm",
   "author": {
     "name": "Shinnosuke Watanabe",
     "url": "https://github.com/shinnn"
@@ -34,18 +34,15 @@
   "deprecated": false,
   "description": "A list of SPDX license identifiers",
   "devDependencies": {
-    "@shinnn/eslint-config-node": "^5.0.0",
-    "chalk": "^2.3.0",
-    "eslint": "^4.16.0",
+    "@shinnn/eslint-config": "^6.8.7",
+    "chalk": "^2.4.1",
+    "eslint": "^5.10.0",
     "get-spdx-license-ids": "^2.1.0",
-    "log-symbols": "^2.2.0",
-    "loud-rejection": "^1.6.0",
-    "rmfr": "^2.0.0-3",
-    "tape": "^4.8.0",
-    "write-file-atomically": "^1.0.0"
+    "rmfr": "^2.0.0",
+    "tape": "^4.9.1"
   },
   "eslintConfig": {
-    "extends": "@shinnn/node"
+    "extends": "@shinnn"
   },
   "files": [
     "deprecated.json",
@@ -71,8 +68,8 @@
   },
   "scripts": {
     "build": "node build.js",
-    "pretest": "eslint --fix --format=codeframe .",
+    "pretest": "eslint .",
     "test": "node test.js"
   },
-  "version": "3.0.0"
+  "version": "3.0.3"
 }
