@@ -3,7 +3,7 @@ import { FormGroup, FormControl, Validators, ValidationErrors } from '@angular/f
 import { AuthService } from 'app/shared/services/auth.service';
 import { Subscription } from 'rxjs';
 import { Router, ActivatedRoute, Params } from '@angular/router';
-// import { ErrorService } from 'app/shared/classes/error.service';
+import {MatSnackBar} from '@angular/material';
 
 @Component({
   selector: 'app-login',
@@ -13,36 +13,32 @@ import { Router, ActivatedRoute, Params } from '@angular/router';
 export class LoginComponent implements OnInit, OnDestroy {
   form: FormGroup;
   aSub: Subscription;
-  errorMessage: String;
 
   constructor(private auth: AuthService,
               private router: Router,
               private route: ActivatedRoute,
+              public snackBar: MatSnackBar) { }
 
-              // private errorService: ErrorService
-              ) { }
-
-  /** Валидатор пароля */
+  /** Password validator */
   private passwordValidator(control: FormControl): ValidationErrors {
     const value = control.value;
-    /** Проверка на содержание цифр */
+    /** Check if string has numbers */
     const hasNumber = /[0-9]/.test(value);
-    /** Проверка на содержание заглавных букв */
+    /** Check if string has uppercase letters */
     const hasCapitalLetter = /[A-Z]/.test(value);
-    /** Проверка на содержание прописных букв */
+    /** Check if string has lowercase letter */
     const hasLowercaseLetter = /[a-z]/.test(value);
-    /** Проверка на содержание специальных символов*/
+    /** Check if string has special characters */
     const hasSpecSymbol = /[^A-Za-z0-9]/.test(value);
-    /** Проверка на минимальную длину пароля */
+    /** Check if string has a minimal length of 8 characters */
     // const isLengthValid = value ? value.length > 8 : false;
     const emptyInput = value ? value.length === null : null;
-    //console.log(emptyInput);
-  /** Общая проверка */
+    /** General check */
     const passwordValid = hasNumber && hasCapitalLetter && hasLowercaseLetter && hasSpecSymbol;
     // && isLengthValid;
-    
+
     if (!passwordValid && emptyInput != null) {
-      return { invalidPassword: 'Password don\'t valid' };  
+      return { invalidPassword: 'Password don\'t valid' };
     }
     if (emptyInput == null) {
       return { invalidPassword: 'Password is required' };
@@ -60,47 +56,43 @@ export class LoginComponent implements OnInit, OnDestroy {
         Validators.required,
         Validators.minLength(8),
         this.passwordValidator])
-    })
+    });
 
     this.route.queryParams.subscribe((params: Params) => {
-      if(params['registered']){
-        //Now you can enter in system using your personal data
-
-      } else if(params['accessDenied']){
-        //You need to be authorizen in system for access this link
-
+      if (params['registered']) {
+        this.openSnackBar('Now you can log in using your email and password', null);
+      } else if (params['accessDenied']) {
+        this.openSnackBar('You must log in first!', null);
+      } else if (params['sessionFailed']) {
+        this.openSnackBar('Session has expired. Please, log in again.', null);
+        this.auth.logOut();
+        this.router.navigate(['/login']);
       }
-    })
+    });
   }
-  
-  ngOnDestroy(){
-    if(this.aSub){
-      this.aSub.unsubscribe()
+
+  ngOnDestroy() {
+    if (this.aSub) {
+      this.aSub.unsubscribe();
     }
   }
-  
-  // openSnackBar(message: string, action: string) {
-  //   this.snackBar.open(message, action, {
-  //     duration: 2000,
-  //   });
-  // }
 
-  onSubmit(){
-    this.form.disable()
+  openSnackBar(message, action) {
+    this.snackBar.open(message, action, {
+      duration: 2000,
+    });
+  }
+
+  onSubmit() {
+    this.form.disable();
 
     this.aSub = this.auth.login(this.form.value).subscribe(
       () => this.router.navigate(['/overview']),
       error => {
-        this.errorMessage = error.error.message
-        // this.errorService.snackBar(error.error.message)
-        console.log(this.errorMessage)
-        this.form.enable()
+        this.openSnackBar(error.error.message, null);
+        this.form.enable();
       }
-    )
+    );
   }
-
-  // openSnackBar(){
-  //   this.errorService.openSnackBar()
-  // }
 
 }
