@@ -5,7 +5,11 @@ import {MatToolbarModule} from '@angular/material/toolbar';
 import {MatIconModule} from '@angular/material/icon';
 import {MatCardModule} from '@angular/material/card';
 import {MatInputModule} from '@angular/material/input';
-// import {MatSnackBar} from '@angular/material';
+import {MatSnackBarModule} from '@angular/material/snack-bar';
+import {DragDropModule} from '@angular/cdk/drag-drop';
+import {MatDialogModule} from '@angular/material/dialog';
+import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
+import {MatTooltipModule} from '@angular/material/tooltip';
 
 @NgModule({
   declarations: [],
@@ -15,16 +19,25 @@ import {MatInputModule} from '@angular/material/input';
     MatToolbarModule,
     MatIconModule,
     MatCardModule,
-    MatInputModule
-    // MatSnackBar
+    MatInputModule,
+    MatSnackBarModule,
+    DragDropModule,
+    MatDialogModule,
+    MatProgressSpinnerModule,
+    MatTooltipModule
   ],
   exports: [
     MatButtonModule,
     MatToolbarModule,
     MatIconModule,
     MatCardModule,
-    MatInputModule
-    // MatSnackBar
+    MatInputModule,
+    MatSnackBarModule,
+    DragDropModule,
+    MatDialogModule,
+    MatProgressSpinnerModule,
+    MatTooltipModule
   ]
 })
+
 export class MatModule { }
