@@ -8,24 +8,22 @@ import { AuthService } from '../services/auth.service';
 })
 export class AuthGuard implements CanActivate, CanActivateChild {
   constructor(private auth: AuthService,
-              private router: Router){
-
-              }
+              private router: Router) { }
 
   canActivate( route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> {
-    if(this.auth.isAuthenticated()){
-      return of(true)
+    if (this.auth.isAuthenticated()) {
+      return of(true);
     } else {
       this.router.navigate(['/login'], {
         queryParams: {
           accessDenied: true
         }
-      })
-      return of(false)
+      });
+      return of(false);
     }
   }
 
-  canActivateChild(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean>{
-    return this.canActivate(route, state)
+  canActivateChild(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> {
+    return this.canActivate(route, state);
   }
 }
