@@ -1,4 +1,5 @@
-module.exports ={
-    mongoURI: 'mongodb://nataZ:natasneak19&@ds123584.mlab.com:23584/notebook',
-    jwt: 'dev-jwt'
+if(process.env.NODE_ENV === 'production'){
+    module.exports = require('./keys.prod');
+} else {
+    module.exports = require('./keys.dev');
 }
