@@ -1,5 +1,5 @@
-const app = require('./app')
-const port = process.env.PORT | 5000
+const app = require('./app');
+const port = process.env.PORT | 5000;
 
 app.listen(port, () => console.log(`Server has been started on port ${port}`)
-)
+);
