@@ -1,11 +1,18 @@
-const express = require('express')
-const controller = require('../controllers/auth')
-const router = express.Router()
+const express = require('express');
+const passport = require('passport');
+const controller = require('../controllers/auth');
+const upload = require('../middleware/upload');
+const router = express.Router();
 
 // localhost:5000/api/auth/login
-router.post('/login', controller.login)
+router.post('/login', controller.login);
 
 // localhost:5000/api/auth/register
-router.post('/register', controller.register)
+router.post('/register', upload.single('image'), controller.register);
 
-module.exports = router
+//localhost:5000/api/auth/profile
+router.get('/profile', passport.authenticate('jwt', {session: false}), controller.getProfile);
+router.put('/profile', passport.authenticate('jwt', {session: false}), upload.single('image'), controller.editProfile);
+
+
+module.exports = router;
