@@ -1,8 +1,8 @@
-const express = require('express')
-const controller = require('../controllers/profile')
-const router = express.Router()
+const express = require('express');
+const controller = require('../controllers/profile');
+const router = express.Router();
 
-router.get('/', controller.getProfileByUserId)
-router.patch('/', controller.getProfileByUserId)
+router.get('/', controller.getProfileByUserId);
+router.patch('/', controller.getProfileByUserId);
 
-module.exports = router
+module.exports = router;
