@@ -1,8 +1,9 @@
-const express = require('express')
-const controller = require('../controllers/user')
-const router = express.Router()
+const express = require('express');
+const passport = require('passport');
+const controller = require('../controllers/user');
+const router = express.Router();
 
-router.get('/:id', controller.showUser)
-router.delete('/:id', controller.deleteUser)
+router.get('/:id', passport.authenticate('jwt', {session: false}), controller.showUser);
+router.delete('/:id', passport.authenticate('jwt', {session: false}), controller.deleteUser);
 
-module.exports = router
+module.exports = router;
