@@ -1,5 +1,5 @@
 module.exports = (res, error) =>{
-   res.staus(500).json({
+   res.status(500).json({
       success: false,
       message: error.message ? error.message : error
    })
